+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2025-01-02
+### Added
+* Added document minification as a feature before document translation, to 
+  allow translation of large docx or pptx files. For more info check the README.
 
 ## [1.11.0] - 2024-11-15
 ### Added